deployment/requirements.txt
streamlit
tensorflow
numpy
Pillow
pickle5
transformers
datasets
pandas
matplotlib